[🚀 사이클2 - 미션 (예약 대기 승인)] 워넬 미션 제출합니다. (#434)
* docs: 기능 요구 사항 정리

* feat: 예약 삭제 시, 첫 번째 예약 대기 인원 자동 전환 기능 추가

순서
- 사용자 예약 삭제 시, 같은 정보로 예약 대기 존재하는지 탐색
- 예약 대기가 존재할 경우 첫 번째 예약 대기 정보 반환
- 첫 번째 예약 대기 이름을 예약으로 update
- 예약 대기 삭제

* docs: 기능 요구 사항 문서 정리

* refactor: 예약 및 예약 대기 Service 조립 구조로 변경

- 통합 쿼리 제거
- MyReservation record에서 class로 변경
- findAllByName() 수정

* refactor: resource, status 객체 내부로 이동

- 예약과 예약 대기에 대한 객체 정보를 객체 내부 생성자에서 생성할 때 부여할 수 있도록 변경

* refactor: 예약 및 예약 대기 조회 로직 수정

- 기존 매번 selectById(themeId)로 Theme를 조회했기 때문에 중복 조회를 제거하기 위해서 Map과 computeIfAbsent 메서드 도입

* refactor: 예약 매개변수 수정 및 검증 책임 변경

- validateExistsWaiting 및 autoApprove 매개변수를 Reservation 객체로 변경
- validateDateTime 검증을 Reservation내부로 책임 변경

* refactor: 예약 승인 책임 도메인 내부로 변경

- 기존 ReservationService의 AutoApprove 메서드에서 DAO 호출 순서로 승인 흐름을 부여했다. 예약 객체가 “이 대기자를 예약자로 승인한다”는 의미의 메서드를 갖고, 서비스는 조회와 저장 흐름을 조율하는 방식으로 바꿔 더 객체지향적이게 변경되었다.

* fix: 예약 조회, 예약 대기 조회에 대한 row lock 추가

deleteByIdIfNameMatches 내
- 취소 중 예약 대기 신청
- 첫 예약 대기 조회 후 예약 대기 삭제
등과 같은 정합성 문제를 해결하기 위해 ForUpdate 메서드 생성

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,7 +1,11 @@
 package roomescape.reservation.service;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.exception.ErrorCode;
@@ -9,20 +13,26 @@ import roomescape.exception.RoomescapeException;
 import roomescape.reservation.MyReservation;
 import roomescape.reservation.Reservation;
 import roomescape.reservation.dao.ReservationDao;
+import roomescape.theme.Theme;
 import roomescape.theme.dao.ThemeDao;
 import roomescape.time.ReservationTime;
 import roomescape.time.dao.TimeDao;
+import roomescape.waiting.ReservationWaiting;
+import roomescape.waiting.dao.ReservationWaitingDao;
 
 @Service
 public class ReservationService {
     private final ReservationDao reservationDao;
     private final ThemeDao themeDao;
     private final TimeDao timeDao;
+    private final ReservationWaitingDao reservationWaitingDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, TimeDao timeDao) {
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, TimeDao timeDao,
+                              ReservationWaitingDao reservationWaitingDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.timeDao = timeDao;
+        this.reservationWaitingDao = reservationWaitingDao;
     }
 
     public List<Reservation> findAll() {
@@ -34,23 +44,40 @@ public class ReservationService {
                 .orElseThrow(() -> new RoomescapeException(ErrorCode.RESERVATION_NOT_FOUND));
     }
 
+    @Transactional(readOnly = true)
     public List<MyReservation> findAllByName(String name) {
-        return reservationDao.selectAllCombinedByName(name);
+        List<MyReservation> myReservations = new ArrayList<>();
+        Map<Long, Theme> themes = new HashMap<>();
+
+        List<Reservation> reservations = reservationDao.selectByName(name);
+        for (Reservation reservation : reservations) {
+            Theme theme = getThemeById(reservation.getThemeId(), themes);
+            myReservations.add(new MyReservation(reservation, theme));
+        }
+
+        List<ReservationWaiting> reservationWaitings = reservationWaitingDao.selectByName(name);
+        for (ReservationWaiting reservationWaiting : reservationWaitings) {
+            Theme theme = getThemeById(reservationWaiting.getThemeId(), themes);
+            myReservations.add(new MyReservation(reservationWaiting, theme));
+        }
+
+        return myReservations;
     }
 
     @Transactional
     public Reservation add(String name, Long themeId, LocalDate date, Long timeId) {
+        validateThemeExists(themeId);
         ReservationTime time = timeDao.selectById(timeId)
                 .orElseThrow(() -> new RoomescapeException(ErrorCode.RESERVATION_TIME_NOT_FOUND));
-        validateDateTime(date, time, ErrorCode.PAST_RESERVATION);
-        validateThemeExists(themeId);
+
+        Reservation newReservation = new Reservation(name, themeId, date, time);
+        newReservation.validateDateTime(date, time, ErrorCode.CANNOT_RESERVE_PAST_DATETIME);
 
         List<Reservation> reservedList = reservationDao.selectByThemeIdAndDate(themeId, date);
         for (Reservation reserved : reservedList) {
             validateReserved(timeId, reserved.getTime());
         }
 
-        Reservation newReservation = new Reservation(name, themeId, date, time);
         return reservationDao.insert(newReservation);
     }
 
@@ -63,7 +90,7 @@ public class ReservationService {
 
         ReservationTime time = timeDao.selectById(timeId)
                 .orElseThrow(() -> new RoomescapeException(ErrorCode.RESERVATION_TIME_NOT_FOUND));
-        validateDateTime(date, time, ErrorCode.PAST_RESERVATION);
+        originReservation.validateDateTime(date, time, ErrorCode.CANNOT_RESERVE_PAST_DATETIME);
         validateThemeExists(themeId);
 
         List<Reservation> reservedList = reservationDao.selectByThemeIdAndDate(themeId, date);
@@ -79,27 +106,35 @@ public class ReservationService {
         reservationDao.deleteById(id);
     }
 
+    @Transactional
     public void deleteByIdIfNameMatches(Long id, String name) {
-        Reservation originReservation = reservationDao.selectById(id)
+        Reservation reservation = reservationDao.selectByIdForUpdate(id)
                 .orElseThrow(() -> new RoomescapeException(ErrorCode.RESERVATION_NOT_FOUND));
 
-        originReservation.validateSameName(name, ErrorCode.CANNOT_DELETE_OTHER_RESERVATION);
-
-        validateDateTime(originReservation.getDate(), originReservation.getTime(),
+        reservation.validateSameName(name, ErrorCode.CANNOT_DELETE_OTHER_RESERVATION);
+        reservation.validateDateTime(reservation.getDate(), reservation.getTime(),
                 ErrorCode.CANNOT_DELETE_PAST_RESERVATION);
 
-        reservationDao.deleteById(id);
+        Optional<ReservationWaiting> firstWaiting = reservationWaitingDao.selectFirstWaitingForUpdate(reservation);
+        if (firstWaiting.isEmpty()) {
+            reservationDao.deleteById(id);
+            return;
+        }
+
+        Reservation approvedReservation = reservation.approve(firstWaiting.get());
+        reservationDao.updateNameByThemeIdAndDateAndTimeId(approvedReservation)
+                .orElseThrow(() -> new RoomescapeException(ErrorCode.RESERVATION_NOT_FOUND));
+        reservationWaitingDao.deleteById(firstWaiting.get().getId());
+    }
+
+    private Theme getThemeById(Long themeId, Map<Long, Theme> themes) {
+        return themes.computeIfAbsent(themeId, id -> themeDao.selectById(themeId)
+                .orElseThrow(() -> new RoomescapeException(ErrorCode.THEME_NOT_FOUND)));
     }
 
     private void validateReserved(Long timeId, ReservationTime reservedTime) {
         if (timeId.equals(reservedTime.getId())) {
             throw new RoomescapeException(ErrorCode.RESERVATION_ALREADY_EXISTS);
-        }
-    }
-
-    private void validateDateTime(LocalDate date, ReservationTime time, ErrorCode errorCode) {
-        if (time.isBeforeDateTime(date, time)) {
-            throw new RoomescapeException(errorCode);
         }
     }
